backend/requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
websockets>=12.0

# Vision & Face Recognition
opencv-python>=4.8.0
mss>=9.0.1
deepface>=0.0.75
tensorflow>=2.13.0
numpy>=1.26.0
Pillow>=10.2.0

# AI Services
google-generativeai>=0.7.0
elevenlabs>=1.0.0
cloudinary>=1.40.0

# Speech (Conversation Copilot)
openai-whisper>=20231117
sounddevice>=0.4.6

# Audio playback
pygame>=2.5.0

# Utilities
requests>=2.31.0
aiofiles>=23.2.1
